+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM rust:1.67.1-slim-bullseye AS builder
+FROM rust:1.69.0-slim-bullseye AS builder
 
 ARG LINDERA_VERSION
 
@@ -16,9 +16,9 @@ RUN set -ex \
 
 COPY . .
 
-RUN rustup component add rustfmt --toolchain 1.67.1-x86_64-unknown-linux-gnu
+RUN rustup component add rustfmt --toolchain 1.69.0-x86_64-unknown-linux-gnu
 
-RUN cargo build --release --features=ipadic,unidic,ko-dic,cc-cedict
+RUN cargo build --release --features=ipadic,ipadic-filter,unidic,unidic-filter,ko-dic,ko-dic-filter,cc-cedict,cc-cedict-filter
 
 FROM debian:bullseye-slim
