+++ Makefile
@@ -12,6 +12,7 @@ LINDERA_CC_CEDICT_VERSION ?= $(shell cargo metadata --no-deps --format-version=1
 LINDERA_DICTIONARY_VERSION ?= $(shell cargo metadata --no-deps --format-version=1 | jq -r '.packages[] | select(.name=="lindera-dictionary") | .version')
 LINDERA_TOKENIZER_VERSION ?= $(shell cargo metadata --no-deps --format-version=1 | jq -r '.packages[] | select(.name=="lindera-tokenizer") | .version')
 LINDERA_FILTER_VERSION ?= $(shell cargo metadata --no-deps --format-version=1 | jq -r '.packages[] | select(.name=="lindera-filter") | .version')
+LINDERA_ANALYZER_VERSION ?= $(shell cargo metadata --no-deps --format-version=1 | jq -r '.packages[] | select(.name=="lindera-analyzer") | .version')
 LINDERA_VERSION ?= $(shell cargo metadata --no-deps --format-version=1 | jq -r '.packages[] | select(.name=="lindera") | .version')
 LINDERA_CLI_VERSION ?= $(shell cargo metadata --no-deps --format-version=1 | jq -r '.packages[] | select(.name=="lindera-cli") | .version')
 
@@ -88,6 +89,10 @@ ifeq ($(shell curl -s -XGET https://crates.io/api/v1/crates/lindera-tokenizer | 
 endif
 ifeq ($(shell curl -s -XGET https://crates.io/api/v1/crates/lindera-filter | jq -r '.versions[].num' | grep $(LINDERA_FILTER_VERSION)),)
 	(cd lindera-filter && cargo package && cargo publish)
+	sleep 10
+endif
+ifeq ($(shell curl -s -XGET https://crates.io/api/v1/crates/lindera-analyzer | jq -r '.versions[].num' | grep $(LINDERA_ANALYZER_VERSION)),)
+	(cd lindera-analyzer && cargo package && cargo publish)
 	sleep 10
 endif
 ifeq ($(shell curl -s -XGET https://crates.io/api/v1/crates/lindera | jq -r '.versions[].num' | grep $(LINDERA_VERSION)),)
