+++ lindera-cc-cedict/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-cc-cedict"
-version = "0.23.0"
+version = "0.24.0"
 edition = "2021"
 description = "A Japanese morphological dictionary for CC-CEDICT."
 documentation = "https://docs.rs/lindera-cc-cedict"
@@ -20,12 +20,12 @@ bincode = "1.3.3"
 byteorder = "1.4.3"
 once_cell = "1.17.0"
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-decompress = { version = "0.23.0", path = "../lindera-decompress", optional = true }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-decompress = { version = "0.24.0", path = "../lindera-decompress", optional = true }
 
 [build-dependencies]
 encoding = { version = "0.2.33", optional = true }
 zip = { version = "0.6.3", optional = true }
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-cc-cedict-builder = { version = "0.23.0", path = "../lindera-cc-cedict-builder"}
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-cc-cedict-builder = { version = "0.24.0", path = "../lindera-cc-cedict-builder"}
