+++ lindera-cli/src/main.rs
@@ -21,16 +21,17 @@ use clap::{Parser, Subcommand};
     feature = "ko-dic-filter",
     feature = "cc-cedict-filter",
 ))]
-use lindera::analyzer::Analyzer;
+use lindera_analyzer::analyzer::Analyzer;
 
-use lindera::{
+use lindera_core::{
     error::{LinderaError, LinderaErrorKind},
     mode::Mode,
-    DictionaryKind, FilteredToken, LinderaResult,
+    LinderaResult,
 };
 use lindera_dictionary::{
-    build_dictionary, build_user_dictionary, DictionaryConfig, UserDictionaryConfig,
+    build_dictionary, build_user_dictionary, DictionaryConfig, DictionaryKind, UserDictionaryConfig,
 };
+use lindera_filter::token::FilteredToken;
 use lindera_tokenizer::tokenizer::{Tokenizer, TokenizerConfig, CONTAINED_DICTIONARIES};
 
 #[derive(Debug, Parser)]
