+++ lindera-core/src/lib.rs
@@ -4,9 +4,9 @@ pub mod dictionary;
 pub mod dictionary_builder;
 pub mod error;
 pub mod file_util;
+pub mod mode;
 pub mod prefix_dict;
 pub mod unknown_dictionary;
-pub mod user_dictionary;
 pub mod viterbi;
 pub mod word_entry;
