+++ lindera-decompress/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-decompress"
-version = "0.23.0"
+version = "0.24.0"
 edition = "2021"
 description = "A morphological analysis library."
 documentation = "https://docs.rs/lindera-decompress"
