+++ lindera-dictionary/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-dictionary"
-version = "0.23.0"
+version = "0.24.0"
 edition = "2021"
 description = "A Japanese morphological dictionary."
 documentation = "https://docs.rs/lindera-dictionary"
@@ -28,12 +28,12 @@ bincode = "1.3.3"
 byteorder = "1.4.3"
 serde = { version="1.0.152", features = ["derive"] }
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-cc-cedict = { version = "0.23.0", path = "../lindera-cc-cedict", optional = true }
-lindera-cc-cedict-builder = { version = "0.23.0", path = "../lindera-cc-cedict-builder" }
-lindera-ipadic = { version = "0.23.0", path = "../lindera-ipadic", optional = true }
-lindera-ipadic-builder = { version = "0.23.0", path = "../lindera-ipadic-builder" }
-lindera-ko-dic = { version = "0.23.0", path = "../lindera-ko-dic", optional = true }
-lindera-ko-dic-builder = { version = "0.23.0", path = "../lindera-ko-dic-builder" }
-lindera-unidic = { version = "0.23.0", path = "../lindera-unidic", optional = true }
-lindera-unidic-builder = { version = "0.23.0", path = "../lindera-unidic-builder" }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-cc-cedict = { version = "0.24.0", path = "../lindera-cc-cedict", optional = true }
+lindera-cc-cedict-builder = { version = "0.24.0", path = "../lindera-cc-cedict-builder" }
+lindera-ipadic = { version = "0.24.0", path = "../lindera-ipadic", optional = true }
+lindera-ipadic-builder = { version = "0.24.0", path = "../lindera-ipadic-builder" }
+lindera-ko-dic = { version = "0.24.0", path = "../lindera-ko-dic", optional = true }
+lindera-ko-dic-builder = { version = "0.24.0", path = "../lindera-ko-dic-builder" }
+lindera-unidic = { version = "0.24.0", path = "../lindera-unidic", optional = true }
+lindera-unidic-builder = { version = "0.24.0", path = "../lindera-unidic-builder" }
