+++ lindera-filter/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-filter"
-version = "0.23.1"
+version = "0.24.0"
 edition = "2021"
 description = "Character and token filters for Lindera."
 documentation = "https://docs.rs/lindera-filter"
@@ -32,8 +32,8 @@ unicode-normalization = "0.1.22"
 unicode-segmentation = "1.10.1"
 yada = "0.5.0"
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-dictionary = { version = "0.23.0", path = "../lindera-dictionary" }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-dictionary = { version = "0.24.0", path = "../lindera-dictionary" }
 
 [dev-dependencies]
 once_cell = "1.17.0"
