+++ lindera-ipadic-builder/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-ipadic-builder"
-version = "0.23.0"
+version = "0.24.0"
 edition = "2021"
 description = "A Japanese morphological dictionary builder for IPADIC."
 documentation = "https://docs.rs/lindera-ipadic-builder"
@@ -27,6 +27,6 @@ log = "0.4.17"
 serde = "1.0.152"
 yada = "0.5.0"
 
-lindera-compress = { version = "0.23.0", path = "../lindera-compress", optional = true }
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-decompress = { version = "0.23.0", path = "../lindera-decompress" }
+lindera-compress = { version = "0.24.0", path = "../lindera-compress", optional = true }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-decompress = { version = "0.24.0", path = "../lindera-decompress" }
