+++ lindera-ko-dic-builder/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-ko-dic-builder"
-version = "0.23.0"
+version = "0.24.0"
 edition = "2021"
 description = "A Korean morphological dictionary builder for ko-dic."
 documentation = "https://docs.rs/lindera-ko-dic-builder"
@@ -25,6 +25,6 @@ glob = "0.3.1"
 log = "0.4.17"
 yada = "0.5.0"
 
-lindera-compress = { version = "0.23.0", path = "../lindera-compress", optional = true }
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-decompress = { version = "0.23.0", path = "../lindera-decompress" }
+lindera-compress = { version = "0.24.0", path = "../lindera-compress", optional = true }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-decompress = { version = "0.24.0", path = "../lindera-decompress" }
