+++ lindera-tokenizer/Cargo.toml
@@ -29,8 +29,8 @@ once_cell = "1.17.0"
 serde = {version="1.0.152", features = ["derive"] }
 serde_json = "1.0.91"
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-dictionary = { version = "0.23.0", path = "../lindera-dictionary" }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-dictionary = { version = "0.24.0", path = "../lindera-dictionary" }
 
 [dev-dependencies]
 criterion = { version = "0.4.0", features = ["html_reports"] }
