+++ lindera-unidic/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera-unidic"
-version = "0.23.0"
+version = "0.24.0"
 edition = "2021"
 description = "A Japanese morphological dictionary for UniDic."
 documentation = "https://docs.rs/lindera-unidic"
@@ -20,13 +20,13 @@ bincode = "1.3.3"
 byteorder = "1.4.3"
 once_cell = "1.17.0"
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-decompress = { version = "0.23.0", path = "../lindera-decompress", optional = true }
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-decompress = { version = "0.24.0", path = "../lindera-decompress", optional = true }
 
 [build-dependencies]
 encoding = { version = "0.2.33", optional = true }
 zip = { version = "0.6.3", optional = true }
 ureq = { version = "2.5.0", default-features = false, features = ["tls"], optional = true }
 
-lindera-core = { version = "0.23.0", path = "../lindera-core" }
-lindera-unidic-builder = { version = "0.23.0", path = "../lindera-unidic-builder"}
+lindera-core = { version = "0.24.0", path = "../lindera-core" }
+lindera-unidic-builder = { version = "0.24.0", path = "../lindera-unidic-builder"}
