--- lindera/src/character_filter.rs
@@ -1,14 +1,0 @@
-pub type JapaneseIterationMarkCharacterFilterConfig =
-    lindera_filter::character_filter::japanese_iteration_mark::JapaneseIterationMarkCharacterFilterConfig;
-pub type JapaneseIterationMarkCharacterFilter =
-    lindera_filter::character_filter::japanese_iteration_mark::JapaneseIterationMarkCharacterFilter;
-pub type MappingCharacterFilterConfig =
-    lindera_filter::character_filter::mapping::MappingCharacterFilterConfig;
-pub type MappingCharacterFilter = lindera_filter::character_filter::mapping::MappingCharacterFilter;
-pub type RegexCharacterFilterConfig =
-    lindera_filter::character_filter::regex::RegexCharacterFilterConfig;
-pub type RegexCharacterFilter = lindera_filter::character_filter::regex::RegexCharacterFilter;
-pub type UnicodeNormalizeCharacterFilterConfig =
-    lindera_filter::character_filter::unicode_normalize::UnicodeNormalizeCharacterFilterConfig;
-pub type UnicodeNormalizeCharacterFilter =
-    lindera_filter::character_filter::unicode_normalize::UnicodeNormalizeCharacterFilter;
