--- lindera/src/error.rs
@@ -1,2 +1,0 @@
-pub type LinderaError = lindera_core::error::LinderaError;
-pub type LinderaErrorKind = lindera_core::error::LinderaErrorKind;
