--- lindera/src/mode.rs
@@ -1,4 +1,0 @@
-use lindera_core::viterbi::{Mode as LinderaCoreMode, Penalty as LinderaCorePenalty};
-
-pub type Penalty = LinderaCorePenalty;
-pub type Mode = LinderaCoreMode;
