--- lindera/src/token_filter.rs
@@ -1,54 +1,0 @@
-pub type JapaneseBaseFormTokenFilterConfig =
-    lindera_filter::token_filter::japanese_base_form::JapaneseBaseFormTokenFilterConfig;
-pub type JapaneseBaseFormTokenFilter =
-    lindera_filter::token_filter::japanese_base_form::JapaneseBaseFormTokenFilter;
-pub type JapaneseCompoundWordTokenFilterConfig =
-    lindera_filter::token_filter::japanese_compound_word::JapaneseCompoundWordTokenFilterConfig;
-pub type JapaneseCompoundWordTokenFilter =
-    lindera_filter::token_filter::japanese_compound_word::JapaneseCompoundWordTokenFilter;
-pub type JapaneseKanaTokenFilterConfig =
-    lindera_filter::token_filter::japanese_kana::JapaneseKanaTokenFilterConfig;
-pub type JapaneseKanaTokenFilter =
-    lindera_filter::token_filter::japanese_kana::JapaneseKanaTokenFilter;
-pub type JapaneseKatakanaStemTokenFilterConfig =
-    lindera_filter::token_filter::japanese_katakana_stem::JapaneseKatakanaStemTokenFilterConfig;
-pub type JapaneseKatakanaStemTokenFilter =
-    lindera_filter::token_filter::japanese_katakana_stem::JapaneseKatakanaStemTokenFilter;
-pub type JapaneseKeepTagsTokenFilterConfig =
-    lindera_filter::token_filter::japanese_keep_tags::JapaneseKeepTagsTokenFilterConfig;
-pub type JapaneseKeepTagsTokenFilter =
-    lindera_filter::token_filter::japanese_keep_tags::JapaneseKeepTagsTokenFilter;
-pub type JapaneseNumberTokenFilterConfig =
-    lindera_filter::token_filter::japanese_number::JapaneseNumberTokenFilterConfig;
-pub type JapaneseNumberTokenFilter =
-    lindera_filter::token_filter::japanese_number::JapaneseNumberTokenFilter;
-pub type JapaneseReadingFormTokenFilterConfig =
-    lindera_filter::token_filter::japanese_reading_form::JapaneseReadingFormTokenFilterConfig;
-pub type JapaneseReadingFormTokenFilter =
-    lindera_filter::token_filter::japanese_reading_form::JapaneseReadingFormTokenFilter;
-pub type JapaneseStopTagsTokenFilterConfig =
-    lindera_filter::token_filter::japanese_stop_tags::JapaneseStopTagsTokenFilterConfig;
-pub type JapaneseStopTagsTokenFilter =
-    lindera_filter::token_filter::japanese_stop_tags::JapaneseStopTagsTokenFilter;
-pub type KeepWordsTokenFilterConfig =
-    lindera_filter::token_filter::keep_words::KeepWordsTokenFilterConfig;
-pub type KeepWordsTokenFilter = lindera_filter::token_filter::keep_words::KeepWordsTokenFilter;
-pub type KoreanKeepTagsTokenFilterConfig =
-    lindera_filter::token_filter::korean_keep_tags::KoreanKeepTagsTokenFilterConfig;
-pub type KoreanKeepTagsTokenFilter =
-    lindera_filter::token_filter::korean_keep_tags::KoreanKeepTagsTokenFilter;
-pub type KoreanReadingFormTokenFilter =
-    lindera_filter::token_filter::korean_reading_form::KoreanReadingFormTokenFilter;
-pub type KoreanStopTagsTokenFilterConfig =
-    lindera_filter::token_filter::korean_stop_tags::KoreanStopTagsTokenFilterConfig;
-pub type KoreanStopTagsTokenFilter =
-    lindera_filter::token_filter::korean_stop_tags::KoreanStopTagsTokenFilter;
-pub type LengthTokenFilterConfig = lindera_filter::token_filter::length::LengthTokenFilterConfig;
-pub type LengthTokenFilter = lindera_filter::token_filter::length::LengthTokenFilter;
-pub type LowercaseTokenFilter = lindera_filter::token_filter::lowercase::LowercaseTokenFilter;
-pub type MappingTokenFilterConfig = lindera_filter::token_filter::mapping::MappingTokenFilterConfig;
-pub type MappingTokenFilter = lindera_filter::token_filter::mapping::MappingTokenFilter;
-pub type StopWordsTokenFilterConfig =
-    lindera_filter::token_filter::stop_words::StopWordsTokenFilterConfig;
-pub type StopWordsTokenFilter = lindera_filter::token_filter::stop_words::StopWordsTokenFilter;
-pub type UppercaseTokenFilter = lindera_filter::token_filter::uppercase::UppercaseTokenFilter;
